[-126] refactoring collateral.
