--- src/components/ContentUser/Collateral/Collateral.module.css
@@ -1,4 +1,0 @@
-.item {
-  display: inline-block;
-  padding: 3px;
-}
